refactor(dispatcher): decouple pipeline validation stages and implement match/case (Phase 2, #979)
